add install go option to first window
The version of Go shipped with Debian 9 (1.7) does not have support for
Go modules, which are strongly recommended for any new projects. A
backport of Go 1.11 does exist for stretch, but generally Go libraries
only support the two most recent versions of Go, so it seems easier to
use the official distribution.

This patch adds a "Install/Upgrade Go" option to the first window in
`fsadapt`. The with this selected, if Go is not installed to
/usr/local/go, or a more recent version is available, the latest
distribution of Go for the current OS is downloaded and installed to
/usr/local/go.